Time: 8 ms (34.91%), Space: 8.4 MB (88.09%) - LeetHub

+++ 804-unique-morse-code-words/804-unique-morse-code-words.cpp
@@ -10,6 +10,6 @@ public:
             }
             m.insert(temp);
         }
-        return (int)m.size();
+        return m.size();
     }
 };
